+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentGetOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentPutOperation`
+01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentUrlEncodedPutOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentPostOperation`
 01. :gift: Added class `dev.orne.http.client.AbstractStatusDependentUrlEncodedPostOperation`
 01. :gift: Added class `dev.orne.http.client.VoidResponseHandler`